requirements.txt
fastapi
pyjwt
bcrypt
passlib
sqlalchemy